csv read write operation optimized, integration test improved

+++ domain/integration/services/DataIntegrationConnectionDatabaseService.py
@@ -1,17 +1,11 @@
 from injector import inject
-from domain.connection.services.ConnectionService import ConnectionService
-from domain.integration.services.DataIntegrationColumnService import DataIntegrationColumnService
 from infrastructor.data.DatabaseSessionManager import DatabaseSessionManager
 from infrastructor.data.Repository import Repository
 from infrastructor.dependency.scopes import IScoped
-from infrastructor.exceptions.OperationalException import OperationalException
 from models.dao.integration import DataIntegrationConnectionDatabase
 from models.dao.integration.DataIntegrationConnection import DataIntegrationConnection
-from models.enums import ConnectionTypes
 from models.viewmodels.integration.CreateDataIntegrationConnectionDatabaseModel import \
     CreateDataIntegrationConnectionDatabaseModel
-from models.viewmodels.integration.CreateDataIntegrationModel import CreateDataIntegrationModel
-from models.dao.integration.DataIntegration import DataIntegration
 
 
 class DataIntegrationConnectionDatabaseService(IScoped):
@@ -24,7 +18,14 @@ class DataIntegrationConnectionDatabaseService(IScoped):
             Repository[DataIntegrationConnectionDatabase](database_session_manager)
 
     #######################################################################################
-    def get_by_data_integration_connection_id(self, data_integration_connection_id: int) -> DataIntegrationConnectionDatabase:
+    def get_by_id(self, id: int) -> DataIntegrationConnectionDatabase:
+        entity = self.data_integration_connection_database_repository.first(IsDeleted=0,
+                                                                            Id=id,
+                                                                            )
+        return entity
+
+    def get_by_data_integration_connection_id(self,
+                                              data_integration_connection_id: int) -> DataIntegrationConnectionDatabase:
         entity = self.data_integration_connection_database_repository.first(IsDeleted=0,
                                                                             DataIntegrationConnectionId=data_integration_connection_id,
                                                                             )
@@ -33,7 +34,6 @@ class DataIntegrationConnectionDatabaseService(IScoped):
     def insert(self,
                data_integration_connection: DataIntegrationConnection,
                data: CreateDataIntegrationConnectionDatabaseModel) -> DataIntegrationConnectionDatabase:
-
         data_integration_connection_database = DataIntegrationConnectionDatabase(Schema=data.Schema,
                                                                                  TableName=data.TableName,
                                                                                  Query=data.Query,
@@ -54,4 +54,6 @@ class DataIntegrationConnectionDatabaseService(IScoped):
         return data_integration_connection_database
 
     def delete(self, id: int):
-        self.data_integration_connection_database_repository.delete_by_id(id)
+        entity = self.get_by_id(id=id)
+        if entity is not None:
+            self.data_integration_connection_database_repository.delete_by_id(id)
